video/uikit: fix -Wfloat-conversino warning

+++ src/video/uikit/SDL_uikitviewcontroller.m
@@ -523,7 +523,7 @@ static void SDLCALL SDL_HideHomeIndicatorHintChanged(void *userdata, const char 
 
     if (self.keyboardHeight) {
         int rectbottom = (int)(self.textInputRect.y + self.textInputRect.h);
-        int keybottom = self.view.bounds.size.height - self.keyboardHeight;
+        int keybottom = (int)(self.view.bounds.size.height - self.keyboardHeight);
         if (keybottom < rectbottom) {
             offset.y = keybottom - rectbottom;
         }
